TATE_NRNT: fixing buffer overflow UART causing file descriptor errors
read_uart used the size of the send buffer instead of the recv buffer to
limit how much data was received, which led to overflow during board
dumps. A few commits ago this led to it causing bad file descriptor
errors, although the bug has been around since increasing the size of
the UART send buffer

+++ hal/comm_manager.c
@@ -125,7 +125,7 @@ int close_uart_comm(int fd) {
     return RETURN_SUCCESS;
 }
 
-int recv_uart_comm(int fd, uint8_t *data, uint16_t *size, uint16_t max_size) {
+int recv_uart_comm(int fd, uint8_t *data, uint32_t *size, int32_t max_size) {
     if (fd < 0)
         return RETURN_ERROR_PARAM;
     PRINT(DEBUG ,"recv %d\n", fd);
